rename compat stories
