+++ apps/perf-test/src/scenarios/IconButton.tsx
