+++ apps/perf-test/src/scenarios/MenuButton.tsx
