+++ apps/vr-tests/src/stories/Button.stories.tsx
@@ -1,646 +1,258 @@
 import * as React from 'react';
 import Screener, { Steps } from 'screener-storybook/src/screener';
 import { storiesOf } from '@storybook/react';
-import { Button } from '@fluentui/react-button';
-import { AddIcon } from '@fluentui/react-icons-mdl2';
-import { TeamsTheme } from '@fluentui/storybook/lib/themes/v8/index';
-import { withCompatThemeProvider } from '@fluentui/storybook';
-import { FabricDecorator, FabricDecoratorFullWidth } from '../utilities/index';
+import { FabricDecorator, FabricDecoratorTall } from '../utilities/index';
+import {
+  DefaultButton,
+  ActionButton,
+  CompoundButton,
+  IconButton,
+  IButtonProps,
+  CommandBarButton,
+} from '@fluentui/react/lib/Button';
 
-storiesOf('Button Next', module)
+const baseProps: IButtonProps = {
+  iconProps: {
+    iconName: 'AddFriend',
+  },
+  children: 'Button',
+  secondaryText: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor',
+};
+
+const commandProps: IButtonProps = {
+  iconProps: { iconName: 'Add' },
+  text: 'Create account',
+  // eslint-disable-next-line no-alert
+  onClick: () => alert('Clicked'),
+  menuProps: {
+    items: [
+      {
+        key: 'emailMessage',
+        text: 'Email message',
+        iconProps: {
+          iconName: 'Mail',
+        },
+      },
+      {
+        key: 'calendarEvent',
+        text: 'Calendar event',
+        iconProps: {
+          iconName: 'Calendar',
+        },
+      },
+    ],
+  },
+};
+
+storiesOf('Button (compat)', module)
   .addDecorator(FabricDecorator)
   .addDecorator(story => (
     <Screener
       steps={new Steps()
         .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
         .hover('.ms-Button')
         .snapshot('hover', { cropTo: '.testWrapper' })
         .mouseDown('.ms-Button')
         .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
         .end()}
     >
       {story()}
     </Screener>
   ))
-  .addStory('Default', () => <Button>Hello, world</Button>)
-  .addStory('Primary', () => <Button primary>Hello, world</Button>)
-  .addStory('Disabled', () => <Button disabled>Hello, world</Button>)
+  .addStory('Root', () => <DefaultButton {...baseProps} />, { rtl: true })
+  .addStory('Disabled', () => <DefaultButton {...baseProps} disabled={true} />)
+  .addStory('Checked', () => <DefaultButton {...baseProps} checked={true} />)
+  .addStory('Primary', () => <DefaultButton {...baseProps} primary={true} />)
   .addStory('Primary Disabled', () => (
-    <Button primary disabled>
-      Hello, world
-    </Button>
+    <DefaultButton {...baseProps} primary={true} disabled={true} />
   ))
-  .addStory('Ghost', () => <Button ghost>Hello, world</Button>)
-  .addStory('Ghost Disabled', () => (
-    <Button disabled ghost>
-      Hello, world
-    </Button>
+  .addStory('Primary Checked', () => <DefaultButton {...baseProps} primary={true} checked={true} />)
+  .addStory('Primary Anchor', () => (
+    <DefaultButton primary href="http://www.bing.com">
+      Button
+    </DefaultButton>
   ))
-  .addStory('Transparent', () => <Button transparent>Hello, world</Button>)
-  .addStory('Transparent Disabled', () => (
-    <Button disabled transparent>
-      Hello, world
-    </Button>
-  ))
-  .addStory('Anchor', () => <Button href="https://www.bing.com">Hello, world</Button>);
+  .addStory('No Icon', () => <DefaultButton>Button</DefaultButton>)
+  .addStory('Icon Only', () => <DefaultButton iconProps={baseProps.iconProps} />);
 
-storiesOf('Button Next - Teams Theme', module)
+storiesOf('Button Action (compat)', module)
   .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider({ theme: TeamsTheme }))
   .addDecorator(story => (
     <Screener
       steps={new Steps()
         .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
         .hover('.ms-Button')
         .snapshot('hover', { cropTo: '.testWrapper' })
         .mouseDown('.ms-Button')
         .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
         .end()}
     >
       {story()}
     </Screener>
   ))
-  .addStory('Default', () => <Button>Hello, world</Button>)
-  .addStory('Primary', () => <Button primary>Hello, world</Button>)
-  .addStory('Disabled', () => <Button disabled>Hello, world</Button>)
-  .addStory('Primary Disabled', () => (
-    <Button primary disabled>
-      Hello, world
-    </Button>
-  ))
-  .addStory('With icon before content', () => <Button icon="X">Hello, world</Button>);
+  .addStory('Root', () => <ActionButton {...baseProps} />, { rtl: true })
+  .addStory('Disabled', () => <ActionButton {...baseProps} disabled={true} />)
+  .addStory('Checked', () => <ActionButton {...baseProps} checked={true} />)
+  .addStory('No Icon', () => <ActionButton>Button</ActionButton>)
+  .addStory('Icon Only', () => <ActionButton iconProps={baseProps.iconProps} />);
 
-storiesOf('Button Next - With icon before content', module)
+storiesOf('Button Compound (compat)', module)
   .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
   .addDecorator(story => (
     <Screener
       steps={new Steps()
         .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
         .hover('.ms-Button')
         .snapshot('hover', { cropTo: '.testWrapper' })
         .mouseDown('.ms-Button')
         .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-
         .end()}
     >
       {story()}
     </Screener>
   ))
-  .addStory('Default', () => <Button icon="X">Hello, world</Button>)
-  .addStory('Primary', () => (
-    <Button primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button disabled icon="X">
-      Hello, world
-    </Button>
-  ))
+  .addStory('Root', () => <CompoundButton {...baseProps} />, { rtl: true })
+  .addStory('Disabled', () => <CompoundButton {...baseProps} disabled={true} />)
+  .addStory('Checked', () => <CompoundButton {...baseProps} checked={true} />)
+  .addStory('Primary', () => <CompoundButton {...baseProps} primary={true} />)
   .addStory('Primary Disabled', () => (
-    <Button primary disabled icon="X">
-      Hello, world
-    </Button>
+    <CompoundButton {...baseProps} primary={true} disabled={true} />
   ))
-  .addStory('Ghost', () => (
-    <Button ghost icon="X">
-      Hello, world
-    </Button>
+  .addStory('Primary Checked', () => (
+    <CompoundButton {...baseProps} primary={true} checked={true} />
   ))
-  .addStory('Ghost Disabled', () => (
-    <Button disabled ghost icon="X">
-      Hello, world
-    </Button>
+  .addStory('No Icon', () => <CompoundButton {...baseProps} iconProps={undefined} />);
+
+storiesOf('Button Command (compat)', module)
+  .addDecorator(story => (
+    <div style={{ display: 'flex', alignItems: 'stretch', height: '40px' }}>{story()}</div>
   ))
-  .addStory('Transparent', () => (
-    <Button transparent icon="X">
-      Hello, world
-    </Button>
+  .addDecorator(FabricDecoratorTall)
+  .addDecorator(story => (
+    <Screener
+      steps={new Steps()
+        .snapshot('default', { cropTo: '.testWrapper' })
+        .hover('.ms-Button')
+        .snapshot('hover', { cropTo: '.testWrapper' })
+        .mouseDown('.ms-Button')
+        .snapshot('pressed', { cropTo: '.testWrapper' })
+        .click('.ms-Button')
+        .hover('.ms-Button')
+        .snapshot('open', { cropTo: '.testWrapper' })
+        .end()}
+    >
+      {story()}
+    </Screener>
   ))
-  .addStory('Transparent Disabled', () => (
-    <Button disabled transparent icon="X">
-      Hello, world
-    </Button>
+  .addStory('Root', () => <CommandBarButton {...commandProps} />, { rtl: true })
+  .addStory('Disabled', () => <CommandBarButton {...commandProps} disabled={true} />)
+  .addStory('Checked', () => <CommandBarButton {...commandProps} checked={true} />);
+
+storiesOf('Button Split (compat)', module)
+  .addDecorator(FabricDecoratorTall)
+  .addDecorator(story => (
+    <Screener
+      steps={new Steps()
+        .snapshot('default', { cropTo: '.testWrapper' })
+        .hover('.ms-Button:nth-child(1)')
+        .snapshot('hover main', { cropTo: '.testWrapper' })
+        .hover('.ms-Button:nth-child(2)')
+        .snapshot('hover split', { cropTo: '.testWrapper' })
+        // .mouseDown('.ms-Button:nth-child(1)')
+        // .snapshot('pressed main', { cropTo: '.testWrapper' })
+        // .hover('.ms-Button') // reset mouseDown
+        // .mouseUp('.ms-Button:nth-child(2)')
+        // .mouseDown('.ms-Button:nth-child(2)')
+        // .snapshot('pressed split', { cropTo: '.testWrapper' })
+        // .click('.ms-Button:nth-child(2)')
+        // .hover('.ms-Button') // move mouse to make click work
+        // .snapshot('open', { cropTo: '.testWrapper' })
+        .end()}
+    >
+      {story()}
+    </Screener>
+  ))
+  .addStory('Root', () => <DefaultButton {...commandProps} split={true} />, { rtl: true })
+  .addStory('Disabled', () => <DefaultButton {...commandProps} disabled={true} split={true} />)
+  .addStory('Default with Primary Action Disabled', () => (
+    <DefaultButton {...commandProps} primaryDisabled={true} split={true} />
+  ))
+  .addStory('Checked', () => <DefaultButton {...commandProps} checked={true} split={true} />)
+  .addStory('Primary', () => <DefaultButton {...commandProps} primary={true} split={true} />)
+  .addStory('Primary Disabled', () => (
+    <DefaultButton {...commandProps} primary={true} disabled={true} split={true} />
+  ))
+  .addStory('Primary with Primary Action Disabled', () => (
+    <DefaultButton {...commandProps} primaryDisabled={true} primary={true} split={true} />
+  ))
+  .addStory('Primary Checked', () => (
+    <DefaultButton {...commandProps} primary={true} checked={true} split={true} />
+  ))
+  .addStory('Command Split', () => <CommandBarButton {...commandProps} split={true} />);
+
+storiesOf('Button Special Scenarios (compat)', module)
+  .addDecorator(FabricDecorator)
+  .addDecorator(story => (
+    <Screener steps={new Steps().snapshot('default', { cropTo: '.testWrapper' }).end()}>
+      {story()}
+    </Screener>
+  ))
+
+  .addStory('primary with placeholder', () => (
+    <div>
+      <DefaultButton {...baseProps} iconProps={{ iconName: '' }} primary={true} />
+      <br />
+      <DefaultButton {...baseProps} iconProps={{ iconName: 'Add' }} primary={true} />
+    </div>
+  ))
+  .addStory('no flex shrink', () => (
+    <div style={{ width: '300px' }}>
+      <DefaultButton
+        {...baseProps}
+        iconProps={{ iconName: 'Add' }}
+        menuIconProps={{}}
+        styles={{ root: { width: '100%' } }}
+      />
+      <DefaultButton
+        {...baseProps}
+        text="This is a much longer string of text in a constrained space"
+        iconProps={{ iconName: 'Add' }}
+        menuIconProps={{}}
+        styles={{ root: { width: '100%' } }}
+      />
+    </div>
   ));
 
-storiesOf('Button Next - With icon after content', module)
+storiesOf('IconButton Scenarios (compat)', module)
   .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
   .addDecorator(story => (
     <Screener
       steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
+        .snapshot('icon', { cropTo: '.testWrapper' })
         .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
+        .snapshot('hover icon', { cropTo: '.testWrapper' })
         .end()}
     >
       {story()}
     </Screener>
   ))
-  .addStory('Default', () => (
-    <Button icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button primary icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button disabled icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button primary disabled icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button ghost icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button disabled ghost icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button transparent icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button disabled transparent icon="X" iconPosition="after">
-      Hello, world
-    </Button>
-  ));
 
-storiesOf('Button Next - Circular', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
+  .addStory('normal icon button', () => (
+    <div>
+      <IconButton iconProps={{ iconName: 'Globe' }} primary={true} />
+    </div>
   ))
-  .addStory('Default', () => (
-    <Button circular icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button circular primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button circular disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button circular primary disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button circular ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button circular disabled ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button circular transparent icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button circular disabled transparent icon="X">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - Icon only', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Default', () => (
-    <Button iconOnly icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button iconOnly primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button iconOnly disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button iconOnly primary disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button iconOnly ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button iconOnly disabled ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button iconOnly transparent icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button iconOnly disabled transparent icon="X">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - Block', module)
-  .addDecorator(FabricDecoratorFullWidth)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Default', () => (
-    <Button block icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button block primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button block disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button block primary disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button block ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button block disabled ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button block transparent icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button block disabled transparent icon="X">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - Inverted', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Default', () => (
-    <Button inverted icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button inverted primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button inverted disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button inverted primary disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button inverted ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button inverted disabled ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button inverted transparent icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button inverted disabled transparent icon="X">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - Loading', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Default', () => (
-    <Button loading icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary', () => (
-    <Button loading primary icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Disabled', () => (
-    <Button loading disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Primary Disabled', () => (
-    <Button loading primary disabled icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost', () => (
-    <Button loading ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Ghost Disabled', () => (
-    <Button loading disabled ghost icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent', () => (
-    <Button loading transparent icon="X">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button loading disabled transparent icon="X">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - Sizes', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Smallest', () => (
-    <Button icon="X" size="smallest">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Smaller', () => (
-    <Button icon="X" size="smaller">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Small', () => (
-    <Button icon="X" size="small">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Large', () => (
-    <Button icon="X" size="large">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Larger', () => (
-    <Button icon="X" size="larger">
-      Hello, world
-    </Button>
-  ))
-  .addStory('Largest', () => (
-    <Button icon="X" size="largest">
-      Hello, world
-    </Button>
-  ));
-
-storiesOf('Button Next - With styled icon from react-icons via tokens', module)
-  .addDecorator(FabricDecorator)
-  .addDecorator(withCompatThemeProvider)
-  .addDecorator(story => (
-    <Screener
-      steps={new Steps()
-        .snapshot('default', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.add('ms-Fabric--isFocusVisible')",
-        )
-        .executeScript("document.getElementsByClassName('ms-Button')[0].focus()")
-        .snapshot('focus', { cropTo: '.testWrapper' })
-        .executeScript(
-          "document.getElementsByClassName('testWrapper')[0].classList.remove('ms-Fabric--isFocusVisible')",
-        )
-        .hover('.ms-Button')
-        .snapshot('hover', { cropTo: '.testWrapper' })
-        .mouseDown('.ms-Button')
-        .snapshot('pressed', { cropTo: '.testWrapper' })
-        .mouseUp('.ms-Button')
-        .end()}
-    >
-      {story()}
-    </Screener>
-  ))
-  .addStory('Default', () => <Button icon={<AddIcon />} tokens={{ iconSize: '30px' }} />)
-  .addStory('Primary', () => <Button primary icon={<AddIcon />} tokens={{ iconSize: '30px' }} />)
-  .addStory('Disabled', () => <Button disabled icon={<AddIcon />} tokens={{ iconSize: '30px' }} />)
-  .addStory('Primary Disabled', () => (
-    <Button primary disabled icon={<AddIcon />} tokens={{ iconSize: '30px' }} />
-  ))
-  .addStory('Ghost', () => <Button ghost icon={<AddIcon />} tokens={{ iconSize: '30px' }} />)
-  .addStory('Ghost Disabled', () => (
-    <Button disabled ghost icon={<AddIcon />} tokens={{ iconSize: '30px' }} />
-  ))
-  .addStory('Transparent', () => (
-    <Button transparent icon={<AddIcon />} tokens={{ iconSize: '30px' }} />
-  ))
-  .addStory('Transparent Disabled', () => (
-    <Button disabled transparent icon={<AddIcon />} tokens={{ iconSize: '30px' }} />
+  .addStory('icon button with menu', () => (
+    <div>
+      <IconButton
+        iconProps={{ iconName: 'Globe' }}
+        primary={true}
+        menuProps={{
+          items: [
+            { key: 'a', text: 'Item 1' },
+            { key: 'b', text: 'Item 2' },
+          ],
+        }}
+      />
+    </div>
   ));
